requirements.txt
openai==1.61.0
streamlit==1.42.0
python-dotenv==1.0.1